requirements.txt
streamlit>=1.32,<2
pandas>=2.0,<3
plotly>=5.18,<6
snowflake-connector-python>=3.10,<6
python-dotenv>=1.0,<2
slack-sdk>=3.27,<4
APScheduler>=3.10,<4
pytz>=2024.1
